fix: 좌표의 형식을 Double로 변경

+++ backend/src/main/java/ssafy/modo/jamkkaebi/common/tmap/dto/RouteQueryDto.java
@@ -7,8 +7,8 @@ import lombok.Data;
 @Builder
 public class RouteQueryDto {
 
-    private Float startX;
-    private Float startY;
-    private Float endX;
-    private Float endY;
+    private Double startX;
+    private Double startY;
+    private Double endX;
+    private Double endY;
 }
